@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
 		<OutputType>WinExe</OutputType>
-		<TargetFramework>net6.0-windows7.0</TargetFramework>
+		<TargetFramework>net7.0-windows7.0</TargetFramework>
 		<UseWPF>true</UseWPF>
 		<Authors>Blake Pell</Authors>
 		<Company></Company>
