@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <Version>2022.8.18.1</Version>
     <Company>Blake Pell</Company>
     <Authors>Blake Pell</Authors>
