@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0-windows7.0</TargetFramework>
+		<TargetFramework>net7.0-windows7.0</TargetFramework>
 		<UseWPF>true</UseWPF>
     <Version>2022.8.18.1</Version>
 		<Authors>Blake Pell</Authors>
