@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <RootNamespace>Avalon.Web</RootNamespace>
     <UserSecretsId>4d382530-b0a6-4577-8de9-44de7f9ebe08</UserSecretsId>
     <LangVersion>latest</LangVersion>
